+++ backend/canisters/local_user_index/api/src/updates/mod.rs
@@ -1,6 +1,7 @@
 pub mod c2c_create_user;
 pub mod c2c_mark_events_migrated_to_stable_memory;
 pub mod c2c_notify_low_balance;
+pub mod c2c_notify_user_events;
 pub mod c2c_notify_user_index_events;
 pub mod c2c_push_wasm_chunk;
 pub mod c2c_upgrade_user_canister_wasm;